Re-visit basic auth protected registry
Using a registry having basic auth enabled the authentication was not
submitted to Clair. This newly introduced error fixes the missing auth
but needs testing with non-basic auth protected registries.

Also maybe the test for the "malformed auth..." string can now be
removed as it does not trigger on registries with basic auth enabled?

Signed-off-by: Knut Ahlers <knut@ahlers.me>