server: add sort option
This adds a '--sort' to the 'server' sub-command. Currently there is
only one sort option, which is by create date. I assumed that his is
useful in general, compared to a random order, so I made this the
default.

This is slightly over engineered, but makes it easier to add additional
sorting options later. One that comes to my mind is sorting by tag (some
kind of natsort/semversort).

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>