Do not pass empty service param to auth challenge
When an auth challenge does not contain a value for service there is no
point in passing an empty value to the challenge endpoint. Moreover, not
passing an empty values makes it easier for a proxy [1] to omit the
service name from a client and set it before forwarding a request.

[1] https://docs.docker.com/registry/recipes/nginx/

Signed-off-by: Athos Ribeiro <athoscr@fedoraproject.org>